Added UI changes to University and updated MySQL connection
Changed UI interfaces of Exam Cell and Career Advisor in University to match new design. Added and updated fields that would be saved in database. Also changed DB connection to start the connection with the database at the beginning of the application. The connection would now exist throughout the application run time.