Database connection updated and Exam Cell CRUD operations added
The database earlier had a session wait_timeout of only 28800 milliseconds which is now updated to 31536000 milliseconds. This will prevent the connection timeout when the user takes a long time to enter/update/delete the data and the respective operations on the database are needed to be performed.
The Exam Cell user can now perform CRUD operations on the student data and also has a search functionality where the user can search for a particular student or a particular group of students. Necessary validations on each and every field are also added.