Added View Applications to Company Enterprise
Recruiting departments in Companies can now view all the applicants applying to their job listings and can also view each applicant or search through applicants. A new status field is also updated when student applies to a particular job listing. This status is then reflected in the view applications panel of the company.