Added function to save accepted job application and basic listing details to database
Accepted job applications along with basic listing details are now saved in the database. This will enable HR role to modify basic details like salary and start date of individual accepted candidates. Further if any candidate goes on hold, the said database entry would be deleted from database so that HR would not be able to view that entry.