Added Rental Lister role with map functionality
A new role that can perform CRUD operations on rental listings was added in the Rental Enterprise. This role can post new listings or view, delete, update previous listings as and when necessary. This role also gets the opportunity to utilize a special map function, wherein this role can directly add map markers of the place whose rent listing it is posting. This map marker from the posting will then be retrieved whenever this listing will be viewed by anyone and will then be viewed on a self updating map.