Added delete function to Company Admin along with some validations
Added salary validation to save and update in job listings panel. Also added delete function in company admin panel. Once company is deleted, the student's access to the job watch portal is now reset in case he was selected by the company for a job. So the student can now apply for other jobs.